refactor(jazz-tools): split graph/mod.rs into compile + execute submodules
graph/mod.rs (3147 LOC) conflated two phases: turning relation IR into
QueryGraph nodes, and settling the graph against storage. They now live
in graph/compile.rs and graph/execute.rs respectively; mod.rs keeps the
type definitions, construction helpers, and result accessors.

No public types renamed; all moves are byte-for-byte. Visibility on a
handful of fields/methods relaxed from private to pub(super) to cross
the file boundary. tests.rs gained explicit Schema/RowPolicyMode imports
that previously rode in via use super::*;.

The realistic_phase1/crud_sustained{,_single_hop} and
observer_write_path benches show no regression (point estimates 1-4%
faster post-split, all p>0.05).
